+++ src/factorial/factorial.h
@@ -1,1 +1,1 @@
-long int factorial(int n);
+long long int factorial(int n);
